requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
antlr4-tools==0.2
anyio==4.9.0
build==1.2.2.post1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
docutils==0.21.2
fastapi==0.109.2
h11==0.16.0
id==1.5.0
idna==3.10
iniconfig==2.1.0
install-jdk==1.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2

